MD-1315 select for update should be inside transaction
select for update locks the row and it can't be updated until the
current transaction ends, but storage.DB() uses a pool of connections
behind the scenes and the update later may use a different connection,
which won't be able to update the row because it is locked. Another
problem is that we don't unlock the row in the end, even if we exit
because of an error for example. Using a transaction should solve the
problem.