MD-709 change password reset for unknown user
when somebody tries to reset password for non-existing user we should
not reveal that user does not exist, as that could be used to check
which users exist and which are not. On the other hand if somebody
genuinely thinks that he has an account, we should inform them that they
do not. Hence we will be sending a message to the email entered about
password reset attempt and that account does not exist.

When somebody goes to password reset confirmation for the unknown user
we act as if the OTP is incorrect and return PermissionDenied error.